% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishdraw.R
\name{custom_params}
\alias{custom_params}
\title{Create and implement a custom JS library.}
\usage{
custom_params(
  path = "./fishdraw.js",
  save = getwd(),
  body_curve_type = 0,
  body_curve_amount = 0.85,
  body_length = 420,
  body_height = 90,
  scale_type = 1,
  scale_scale = 1,
  pattern_type = 3,
  pattern_scale = 1,
  dorsal_texture_type = 1,
  dorsal_type = 0,
  dorsal_length = 100,
  dorsal_start = 8,
  dorsal_end = 27,
  wing_texture_type = 0,
  wing_type = 0,
  wing_start = 6,
  wing_end = 6,
  wing_y = 0.7,
  wing_length = 130,
  wing_width = 10,
  pelvic_start = 9,
  pelvic_end = 14,
  pelvic_length = 85,
  pelvic_type = 0,
  pelvic_texture_type = 0,
  anal_start = 19,
  anal_end = 29,
  anal_length = 50,
  anal_type = 0,
  anal_texture_type = 0,
  tail_type = 0,
  tail_length = 75,
  finlet_type = 0,
  neck_type = 0,
  nose_height = 0,
  mouth_size = 8,
  head_length = 30,
  head_texture_amount = 60,
  has_moustache = 1,
  moustache_length = 10,
  has_beard = 0,
  has_teeth = 1,
  teeth_length = 8,
  teeth_space = 3.5,
  beard_length = 30,
  eye_type = 0,
  eye_size = 10,
  jaw_size = 1,
  jaw_open = 1
)
}
\arguments{
\item{path}{Path for fishdraw.js}

\item{save}{save}

\item{body_curve_type}{\code{numeric} body_curve_type options:0 or 1.}

\item{body_curve_amount}{\code{numeric} body_curve_amount options:0.5,0.85 or 0.98.}

\item{body_length}{\code{numeric} body_length options:200,350 or 420.}

\item{body_height}{\code{numeric} body_height options:45,90 or 150.}

\item{scale_type}{\code{numeric} scale_type options:0,1,2 or 3.}

\item{scale_scale}{\code{numeric} scale_scale options:0.8,1 or 1.5.}

\item{pattern_type}{\code{numeric} pattern_type options:0,1,2,3 or 4.}

\item{pattern_scale}{\code{numeric} pattern_scale options:0.5,1 or 2.}

\item{dorsal_texture_type}{\code{numeric} dorsal_texture_type options:0 or 1.}

\item{dorsal_type}{\code{numeric} dorsal_type options:0 or 1.}

\item{dorsal_length}{\code{numeric} dorsal_length options:30,90 or 180.}

\item{dorsal_start}{\code{numeric} dorsal_start options:7,8,15;11,12,16.}

\item{dorsal_end}{\code{numeric} dorsal_end options:20,27,28;19,21,24.}

\item{wing_texture_type}{\code{numeric} wing_texture_type options:0 or 1.}

\item{wing_type}{\code{numeric} wing_type options:0 or 1.}

\item{wing_start}{\code{numeric} wing_start options:5,6 or 8.}

\item{wing_end}{\code{numeric} wing_end options:5,6 or 8.}

\item{wing_y}{\code{numeric} wing_y options:0.45,0.7 or 0.85.}

\item{wing_length}{\code{numeric} wing_length options:40,130,200;40,150,350.}

\item{wing_width}{\code{numeric} wing_width options:7,10,20;20,30,50.}

\item{pelvic_start}{\code{numeric} pelvic_start options:7,9,11;7,9,12.}

\item{pelvic_end}{\code{numeric} pelvic_end options:13,14,15}

\item{pelvic_length}{\code{numeric} pelvic_length options:0 or 1.}

\item{pelvic_type}{\code{numeric} pelvic_type options:0 or 1.}

\item{pelvic_texture_type}{\code{numeric} pelvic_texture_type options:0 or 1.}

\item{anal_start}{\code{numeric} anal_start options:16,19 or 23.}

\item{anal_end}{\code{numeric} anal_end options:25,29 or 31.}

\item{anal_length}{\code{numeric} anal_length options:20,50 or 80.}

\item{anal_type}{\code{numeric} anal_type options:0 or 1.}

\item{anal_texture_type}{\code{numeric} anal_texture_type options:0 or 1.}

\item{tail_type}{\code{numeric} tail_type options:0,1,2,3,4 or 5.}

\item{tail_length}{\code{numeric} tail_length options:50,75 or 180.}

\item{finlet_type}{\code{numeric} finlet_type options:0,1,2 or 3.}

\item{neck_type}{\code{numeric} neck_type options:0 or 1.}

\item{nose_height}{\code{numeric} nose_height options:-50,0 or 35.}

\item{mouth_size}{\code{numeric} mouth_size options:6,8 or 11.}

\item{head_length}{\code{numeric} head_length options:20,30 or 50.}

\item{head_texture_amount}{\code{numeric} head_texture_amount options:30,60 or 160.}

\item{has_moustache}{\code{numeric} has_moustache options:0,0,0,1.}

\item{moustache_length}{\code{numeric} moustache_length options:10,20,40.}

\item{has_beard}{\code{numeric} has_beard options:0,0,0,0,0,1.}

\item{has_teeth}{\code{numeric} has_teeth options:0,1,1.}

\item{teeth_length}{\code{numeric} teeth_length options:5,8 or 15.}

\item{teeth_space}{\code{numeric} teeth_space options:3,3.5 or 6.}

\item{beard_length}{\code{numeric} beard_length options:20,30 or 50.}

\item{eye_type}{\code{numeric} eye_type options:0 or 1.}

\item{eye_size}{\code{numeric} eye_size options:8,10 or 28.}

\item{jaw_size}{\code{numeric} jaw_size options:0.7,1 or 1.4.}

\item{jaw_open}{\code{numeric} jaw_open options:0 or 1.}
}
\value{
Custom JS library.
}
\description{
Procedurally generated fish drawings via ggplot2 based on fishdraw JavaScript library \url{https://github.com/LingDong-/fishdraw}.
}
\note{
Note that some fish species might not be representable with this system, and passing "bad" params might produce weird results or crash the program.
}
\examples{
\dontrun{
# create and implement a custom JS library
custom_params(path = "./fishdraw.js",save = getwd())

}
}
\author{
Liuyong Ding \email{ly_ding@126.com}
}
