\name{rfishdraw-package}
\alias{rfishdraw-package}
\alias{rfishdraw}
\docType{package}
\title{
Procedurally Generated Fish Drawings Via Javascript
}
\description{
This package is designed for procedurally generated fish drawings via ggplot2 based on fishdraw JavaScript library <https://github.com/LingDong-/fishdraw>.
}
\details{
\packageDESCRIPTION{rfishdraw}
\packageIndices{rfishdraw}
}

\author{Maintainer: \packageMaintainer{rfishdraw}}


\keyword{ package }
