\name{rfishdraw-package}
\alias{rfishdraw-package}
\alias{rfishdraw}
\docType{package}
\title{
Automatically Generated Fish Drawings via JavaScript
}
\description{
This package is designed for programmatically generated fish drawings via ggplot2 based on JavaScript library <https://github.com/LingDong-/fishdraw>.
}
\details{
\packageDESCRIPTION{rfishdraw}
\packageIndices{rfishdraw}
}

\author{Maintainer: \packageMaintainer{rfishdraw}}


\keyword{ package }
